Encode keys as bytewise array + concat buffer for efficiencies
We still need to add a length-prefixed array type to bytewise, but once we
do we can transform this into a valid bytewise encoding. The idea is that
the tail of a length-prefixed array can be absolutely anything, and need
not be escaped.

Logically we can think of non-array keys, which aren't escaped, as
length-prefixed keys of 0 length.